Moved FocusChart and TimeFocusCapped module to shared folder
