+++ src/Parser/Hunter/Marksmanship/CombatLogParser.js
@@ -13,12 +13,8 @@ import CooldownTracker from './Modules/Features/CooldownTracker';
 import AlwaysBeCasting from './Modules/Features/AlwaysBeCasting';
 import VulnerableUptime from './Modules/Features/VulnerableUptime';
 import VulnerableTracker from './Modules/Features/AimedInVulnerableTracker';
-import TimeFocusCapped from './Modules/Features/TimeFocusCapped';
+import TimeFocusCapped from '../Shared/Modules/TimeFocusCapped';
 import VulnerableApplications from "./Modules/Features/VulnerableApplications";
-
-//Focus
-import FocusChart from './Modules/FocusChart/Focus';
-import FocusTracker from './Modules/FocusChart/FocusTracker';
 
 //Tier
 import Tier20_2p from './Modules/Items/Tier20_2p';
@@ -27,6 +23,10 @@ import Tier19_2p from "./Modules/Items/Tier19_2p";
 
 //Spells
 import Trueshot from './Modules/Spells/Trueshot';
+
+//Focus
+import FocusChart from '../Shared/FocusChart/Focus';
+import FocusTracker from '../Shared/FocusChart/FocusTracker';
 
 //Items
 import UllrsFeatherSnowshoes from './Modules/Items/UllrsFeatherSnowshoes';
