+++ src/Parser/Hunter/Shared/FocusChart/Focus.css
@@ -20,3 +20,6 @@
   background-color: #008fff;
   border-color: #008fff;
 }
+.hihi{
+  
+}
