+++ src/Parser/Hunter/Shared/FocusChart/Focus.js
