+++ src/Parser/Hunter/Shared/FocusChart/FocusComponent.js
