+++ src/Parser/Hunter/Shared/FocusChart/FocusTracker.js
